requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Login>=0.6.0
python-dotenv>=0.21.0
PyMySQL>=1.0.2
Werkzeug>=2.0.0
twilio>=7.0.0
gunicorn>=20.0.4
cryptography>=3.4.7
